Fix sql syntax spacing

+++ Wetr-Solution/Wetr.Dal.Ado/AdoUserDao.cs
@@ -33,8 +33,7 @@ namespace Wetr.Dal.Ado
         public async Task<bool> DeleteAsync(int userId)
         {
             return await this.template.ExecuteAsync(
-               @"delete from user" +
-                   "where userId = @userId",
+               @"delete from user where userId = @userId",
                new Parameter("@userId", userId)) == 1;
         }
 
@@ -65,11 +64,11 @@ namespace Wetr.Dal.Ado
 
         public async Task<bool> InsertAsync(User user)
         {
-            // no id since it's set to auto increment
             return await this.template.ExecuteAsync(
-                @"insert into user (firstName, lastName, password, email) VALUES" +
-                    "(@firstName, @lastName, @password, @email)",
+                @"insert into user (userId, firstName, lastName, password, email) VALUES" +
+                    "(@userId, @firstName, @lastName, @password, @email)",
                 new Parameter("@firstName", user.FirstName),
+                new Parameter("@userId", user.UserId),
                 new Parameter("@lastName", user.LastName),
                 new Parameter("@password", user.Password),
                 new Parameter("@email", user.Email)) == 1;
@@ -78,11 +77,11 @@ namespace Wetr.Dal.Ado
         public async Task<bool> UpdateAsync(User user)
         {
             return await this.template.ExecuteAsync(
-                @"update user set" +
-                    "firstName = @firstName" +
-                    "lastName = @lastName" +
-                    "password = @password" +
-                    "email = @email" +
+                @"update user set " +
+                    "firstName = @firstName," +
+                    "lastName = @lastName," +
+                    "password = @password," +
+                    "email = @email " +
                 "where userId = @userId",
                 new Parameter("@firstName", user.FirstName),
                 new Parameter("@lastName", user.LastName),
